+++ .github/workflows/build.yml
@@ -31,12 +31,12 @@ jobs:
           set -eux
           python -m pip install .[test]
 
-          pytest -vv -r ap --cov glue_lab
+          pytest -vv -r ap --cov gluepyter
           jupyter server extension list
-          jupyter server extension list 2>&1 | grep -ie "glue_lab.*OK"
+          jupyter server extension list 2>&1 | grep -ie "gluepyter.*OK"
 
           jupyter labextension list
-          jupyter labextension list 2>&1 | grep -ie "glue-lab.*OK"
+          jupyter labextension list 2>&1 | grep -ie "gluepyter.*OK"
           python -m jupyterlab.browser_check
 
       - name: Package the extension
@@ -45,13 +45,13 @@ jobs:
 
           pip install build
           python -m build
-          pip uninstall -y "glue_lab" jupyterlab
+          pip uninstall -y "gluepyter" jupyterlab
 
       - name: Upload extension packages
         uses: actions/upload-artifact@v3
         with:
           name: extension-artifacts
-          path: dist/glue_lab*
+          path: dist/gluepyter*
           if-no-files-found: error
 
   test_isolated:
@@ -76,14 +76,14 @@ jobs:
           sudo rm -rf $(which node)
           sudo rm -rf $(which node)
 
-          pip install "jupyterlab>=4.0.0,<5" glue_lab*.whl
+          pip install "jupyterlab>=4.0.0,<5" gluepyter*.whl
 
 
           jupyter server extension list
-          jupyter server extension list 2>&1 | grep -ie "glue_lab.*OK"
+          jupyter server extension list 2>&1 | grep -ie "gluepyter.*OK"
 
           jupyter labextension list
-          jupyter labextension list 2>&1 | grep -ie "glue-lab.*OK"
+          jupyter labextension list 2>&1 | grep -ie "gluepyter.*OK"
           python -m jupyterlab.browser_check --no-browser-test
 
   integration-tests:
@@ -109,7 +109,7 @@ jobs:
       - name: Install the extension
         run: |
           set -eux
-          python -m pip install "jupyterlab>=4.0.0,<5" glue_lab*.whl
+          python -m pip install "jupyterlab>=4.0.0,<5" gluepyter*.whl
 
       - name: Install dependencies
         working-directory: ui-tests
@@ -137,7 +137,7 @@ jobs:
         if: always()
         uses: actions/upload-artifact@v3
         with:
-          name: glue_lab-playwright-tests
+          name: gluepyter-playwright-tests
           path: |
             ui-tests/test-results
             ui-tests/playwright-report
