+++ .github/workflows/check-release.yml
@@ -24,5 +24,5 @@ jobs:
       - name: Upload Distributions
         uses: actions/upload-artifact@v3
         with:
-          name: glue_lab-releaser-dist-${{ github.run_number }}
+          name: gluepyter-releaser-dist-${{ github.run_number }}
           path: .jupyter_releaser_checkout/dist
