+++ .gitignore
@@ -7,9 +7,9 @@ node_modules/
 *.egg-info/
 .ipynb_checkpoints
 *.tsbuildinfo
-glue_lab/labextension
+gluepyter/labextension
 # Version file is handled by hatchling
-glue_lab/_version.py
+gluepyter/_version.py
 
 # Integration tests
 ui-tests/test-results/
