+++ binder/postBuild
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-""" perform a development install of glue_lab
+""" perform a development install of gluepyter
 
     On Binder, this will run _after_ the environment has been fully created from
     the environment.yml in this directory.
@@ -40,7 +40,7 @@ _(
     "jupyter",
     "serverextension",
     "enable",
-    "glue_lab",
+    "gluepyter",
 )
 _(
     sys.executable,
@@ -49,7 +49,7 @@ _(
     "server",
     "extension",
     "enable",
-    "glue_lab",
+    "gluepyter",
 )
 
 # verify the environment the extension didn't break anything
@@ -62,5 +62,5 @@ _("jupyter", "server", "extension", "list")
 _("jupyter", "labextension", "list")
 
 
-print("JupyterLab with glue_lab is ready to run with:\n")
+print("JupyterLab with gluepyter is ready to run with:\n")
 print("\tjupyter lab\n")
