+++ conftest.py
@@ -5,4 +5,4 @@ pytest_plugins = ("pytest_jupyter.jupyter_server",)
 
 @pytest.fixture
 def jp_server_config(jp_server_config):
-    return {"ServerApp": {"jpserver_extensions": {"glue_lab": True}}}
+    return {"ServerApp": {"jpserver_extensions": {"gluepyter": True}}}
