--- glue_lab/tests/__init__.py
@@ -1,1 +1,0 @@
-"""Python unit tests for glue_lab."""
