+++ install.json
@@ -1,5 +1,5 @@
 {
   "packageManager": "python",
-  "packageName": "glue_lab",
-  "uninstallInstructions": "Use your Python package manager (pip, conda, etc.) to uninstall the package glue_lab"
+  "packageName": "gluepyter",
+  "uninstallInstructions": "Use your Python package manager (pip, conda, etc.) to uninstall the package gluepyter"
 }
