+++ jupyter-config/nb-config/glue_lab.json
@@ -1,7 +1,7 @@
 {
   "NotebookApp": {
     "nbserver_extensions": {
-      "glue_lab": true
+      "gluepyter": true
     }
   }
 }
