+++ jupyter-config/server-config/glue_lab.json
@@ -1,7 +1,7 @@
 {
   "ServerApp": {
     "jpserver_extensions": {
-      "glue_lab": true
+      "gluepyter": true
     }
   }
 }
