+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "glue-lab",
+  "name": "gluepyter",
   "version": "0.1.0",
   "description": "A JupyterLab extension for glue-viz",
   "keywords": [
@@ -7,9 +7,9 @@
     "jupyterlab",
     "jupyterlab-extension"
   ],
-  "homepage": "https://github.com/QuantStack/glue-lab",
+  "homepage": "https://github.com/QuantStack/gluepyter",
   "bugs": {
-    "url": "https://github.com/QuantStack/glue-lab/issues"
+    "url": "https://github.com/QuantStack/gluepyter/issues"
   },
   "license": "BSD-3-Clause",
   "author": {
@@ -25,7 +25,7 @@
   "style": "style/index.css",
   "repository": {
     "type": "git",
-    "url": "https://github.com/QuantStack/glue-lab.git"
+    "url": "https://github.com/QuantStack/gluepyter.git"
   },
   "scripts": {
     "build": "jlpm build:schema && jlpm build:lib && jlpm build:labextension:dev",
@@ -38,7 +38,7 @@
     "clean": "jlpm clean:lib",
     "clean:lib": "rimraf lib tsconfig.tsbuildinfo",
     "clean:lintcache": "rimraf .eslintcache .stylelintcache",
-    "clean:labextension": "rimraf glue_lab/labextension glue_lab/_version.py",
+    "clean:labextension": "rimraf gluepyter/labextension gluepyter/_version.py",
     "clean:all": "jlpm clean:lib && jlpm clean:labextension && jlpm clean:lintcache",
     "eslint": "jlpm eslint:check --fix",
     "eslint:check": "eslint . --cache --ext .ts,.tsx",
@@ -115,12 +115,12 @@
           "pip"
         ],
         "base": {
-          "name": "glue_lab"
+          "name": "gluepyter"
         }
       }
     },
     "extension": true,
-    "outputDir": "glue_lab/labextension",
+    "outputDir": "gluepyter/labextension",
     "sharedPackages": {
       "yjs": {
         "bundled": false,
