+++ src/__tests__/glue_lab.spec.ts
@@ -2,7 +2,7 @@
  * Example of [Jest](https://jestjs.io/docs/getting-started) unit tests
  */
 
-describe('glue-lab', () => {
+describe('gluepyter', () => {
   it('should be tested', () => {
     expect(1 + 1).toEqual(2);
   });
