+++ ui-tests/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "glue-lab-ui-tests",
+  "name": "gluepyter-ui-tests",
   "version": "1.0.0",
-  "description": "JupyterLab glue-lab Integration Tests",
+  "description": "JupyterLab gluepyter Integration Tests",
   "private": true,
   "scripts": {
     "start": "jupyter lab --config jupyter_server_test_config.py",
